+++ src/Pim/Bundle/CatalogBundle/Doctrine/MongoDBODM/ProductRepository.php
@@ -103,31 +103,13 @@ class ProductRepository extends DocumentRepository implements
     /**
      * {@inheritdoc}
      */
-    public function findOneBy(array $criteria)
+    public function findOneByIdentifier($identifier)
     {
         $pqb = $this->productQueryFactory->create();
         $qb = $pqb->getQueryBuilder();
-
-        foreach ($criteria as $field => $data) {
-            // TODO : fix the calls to this method, no need to pass the attribute object in data, pass only the value
-            if (is_array($data)) {
-                $attribute = $data['attribute'];
-                $field = $attribute->getCode();
-                $data = $data['value'];
-            }
-            $pqb->addFilter($field, '=', $data);
-        }
-
+        $attribute = $this->getIdentifierAttribute();
+        $pqb->addFilter($attribute->getCode(), '=', $identifier);
         $result = $qb->getQuery()->execute();
-
-        if ($result->count() > 1) {
-            throw new \LogicException(
-                sprintf(
-                    'Many products have been found that match criteria:' . "\n" . '%s',
-                    print_r($criteria, true)
-                )
-            );
-        }
 
         return $result->getNext();
     }